+++ src/Config/tests/fixtures/nested.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'id'       => 'hello world',
+    'nested' => [
+        'other' => [
+            'key' => 'value'
+        ]
+    ]
+];